requirements.txt
libvirt-python
six
pytest>=2.3.5
